Move pages to src, deleted header
