+++ frontend/src/components/NavBar.js
@@ -8,6 +8,13 @@ import Logo from '../resources/medshare-logo.png';
 
 const Wrapper = styled.div`
   display: flex;
+  width: 100%;
+  margin: 0 auto;
+  margin-bottom: 1rem;
+`;
+
+const LogoWrapper = styled.div`
+  display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
@@ -50,43 +57,45 @@ const NavBar = () => {
   const [drawer, handleDrawer] = useState(false);
 
   return (
-    <Navbar>
-      <NavbarGroup align={Alignment.LEFT}>
-        <Button className="bp3-minimal" icon="menu" onClick={() => handleDrawer(true)} />
-        <Drawer
-          style={{ background: '#6396B3', color: '#FFF' }}
-          size="255px"
-          isOpen={drawer}
-          onClose={() => handleDrawer(false)}
-          position={Position.LEFT}
-        >
-          <DrawerWrapper>
-            <TopDrawer>
-              <Button classname="bp3-minimal" icon="menu" onClick={() => handleDrawer(false)} />
-            </TopDrawer>
-            <TextWrapper1>
-              <Link to="/" style={{ color: '#FFF', fontSize: '24px', margin: '0.75rem 0.75rem' }}>
-                SEARCH
-              </Link>
-              <Link to="/saved" style={{ color: '#FFF', fontSize: '24px', margin: '0.75rem 0.75rem' }}>
-                FAVORITES
-              </Link>
-              <Link to="/faq" style={{ color: '#FFF', fontSize: '24px', margin: '0.75rem 0.75rem' }}>
-                FAQ
-              </Link>
-            </TextWrapper1>
-            <TextWrapper2>
-              <Link to="/admin" style={{ color: '#FFF', fontSize: '24px' }}>
-                ADMIN PORTAL
-              </Link>
-            </TextWrapper2>
-          </DrawerWrapper>
-        </Drawer>
-      </NavbarGroup>
-      <Wrapper>
-        <img src={Logo} alt="medshare-logo" width="35rem" height="35rem" />
-      </Wrapper>
-    </Navbar>
+    <Wrapper>
+      <Navbar>
+        <NavbarGroup align={Alignment.LEFT}>
+          <Button className="bp3-minimal" icon="menu" onClick={() => handleDrawer(true)} />
+          <Drawer
+            style={{ background: '#6396B3', color: '#FFF' }}
+            size="255px"
+            isOpen={drawer}
+            onClose={() => handleDrawer(false)}
+            position={Position.LEFT}
+          >
+            <DrawerWrapper>
+              <TopDrawer>
+                <Button classname="bp3-minimal" icon="menu" onClick={() => handleDrawer(false)} />
+              </TopDrawer>
+              <TextWrapper1>
+                <Link to="/" style={{ color: '#FFF', fontSize: '24px', margin: '0.75rem 0.75rem' }}>
+                  SEARCH
+                </Link>
+                <Link to="/saved" style={{ color: '#FFF', fontSize: '24px', margin: '0.75rem 0.75rem' }}>
+                  FAVORITES
+                </Link>
+                <Link to="/faq" style={{ color: '#FFF', fontSize: '24px', margin: '0.75rem 0.75rem' }}>
+                  FAQ
+                </Link>
+              </TextWrapper1>
+              <TextWrapper2>
+                <Link to="/admin" style={{ color: '#FFF', fontSize: '24px' }}>
+                  ADMIN PORTAL
+                </Link>
+              </TextWrapper2>
+            </DrawerWrapper>
+          </Drawer>
+        </NavbarGroup>
+        <LogoWrapper>
+          <img src={Logo} alt="medshare-logo" width="35rem" height="35rem" />
+        </LogoWrapper>
+      </Navbar>
+    </Wrapper>
   );
 };
 export default NavBar;
